Creation of the SRS document
The SRS document has been created and formatted after the example sent in the Project 1 Assignment. We will discuss all the requirements and the use-case diagram on September 18 during our meeting. Then, each of us will contribute a fifth of the requirements that we generate at the meeting to fulfill the participation requirement specified in 1.4 of the Project 1 Assignment prompt.