Message Persistence, Permissions
Made messages persistant across activity switches. Permissions are integrated as they should be. Message class now only stores the message content, the timestamp, and whether or not the message was composed by the user. Conversation now holds the phone number of the other phone different from the one running the current application.