Added some classes and refactored some code.
Added the ConversationViewAdapter and ConversationViewHolder classes to be consistent with how we are doing the Activities. Refactored some code in the MessageActivity class to make it more organized (used the extract method). Also added a conversation_box xml file. Still need a lot of work on the UI part of ConversationActivity.