Implemented more code to ConversationActivity class
Implemented the code to change between the ConversationActvity and the MessageActivity in the ConversationActivity class. Also added the code for that to work in the MessageActivity class (I put it in a comment). Minor changes ConversationViewHolder and ConversationViewAdapter. Also got rid of the conversation_box xml since Nathan added the conversation_bubble xml. Will work more on the class later today and tonight.